lab9

+++ operators.c
@@ -1,16 +1,16 @@
 #include "operators.h"
 
-int add(int op1, int op2) {
+double add(int op1, int op2) {
 	return op1+op2;
 }
-int minus(int op1, int op2) {
+double minus(int op1, int op2) {
 	return op1-op2;
 }
-float mul(int op1, int op2) {
+double mul(int op1, int op2) {
 	return op1*op2;
 }
 
-float div(int op1, int op2) {
-	return (float)op1/(float)op2;
+double div(int op1, int op2) {
+	return (double)op1/(double)op2;
 }
